project-samarth-main/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

streamlit==1.38.0

langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.3

chromadb==0.4.24
sentence-transformers==2.5.1
huggingface_hub==0.24.6

pandas==2.2.2
numpy==1.26.4
scipy==1.11.4
requests==2.32.3
python-dotenv==1.0.1
openpyxl==3.1.2

pydantic==1.10.13
typing-extensions==4.12.2
